+++ src/restSpa/configs.nim
@@ -1,6 +1,0 @@
-from std/os import getEnv
-
-const
-  dbHost* = getEnv "dbHost"
-  dbUser* = getEnv "dbUser"
-  dbPass* = getEnv "dbPass"